GetQuestionById was added and project structure was modified
